client/requirements.txt
# requirements.txt - placeholder
httpx>=0.27.0
python-dotenv>=1.0.0